pyproject.toml
[project]
name = "auth_microservice"
description = "auth_microservice"
readme = "README.md"
requires-python = ">=3.12"
version = "1.0.0"
dependencies = [
    "fastapi==0.*",
    "uvicorn==0.*",
    "sqlalchemy==2.*",
    "psycopg2-binary==2.*",
    "sqlmodel==0.*",
    "asyncpg==0.*",
    "loguru==0.*",
    "email-validator==2.*",
    "pytest==8.*",
    "python-dotenv==1.*",
    "pyjwt==2.*",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "black==24.*", 
    "coverage==7.*", 
    "pytest-asyncio>=0.25.2",
    "aiosqlite>=0.20.0",
    "mypy==1.*", 
    "ruff==0.*", 
    "isort==5.*",
]

[tool.black]
line-length = 79
target-version = ["py312"]

[tool.isort]
py_version = 312
profile = "black"
line_length = 79
force_single_line = false
force_sort_within_sections = false
from_first = false
order_by_type = true
no_sections = false
import_heading_firstparty = "Auth Microservice stuff"

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "auto"
addopts = "-ra"

[tool.ruff]
line-length = 79
indent-width = 4
target-version = "py312"

[tool.pyright]
root = ['auth_microservice']

[tool.mypy]
python_version = "3.12"
namespace_packages = true
ignore_missing_imports = true
# Ignore the "Relative import climbs too many namespaces" error
disable_error_code = "misc"
# Exclude files that actually right, but cause the error
exclude = [
    "auth_microservice/src/models/dynamic_db_models.py",
    "auth_microservice/src/models/pkce_models.py",
    "auth_microservice/src/models/dynamic_models.py",
    "auth_microservice/src/models/old_models.py"
]